Abstracted validators reward to a hook
